requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
SQLAlchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
pydantic>=2.7,<3.0
pydantic-settings>=2.2,<3.0
redis==5.0.2
httpx==0.26.0
pytest==8.0.2
pytest-asyncio==0.23.5
aiosqlite==0.20.0
aiogram>=3.7,<4.0
python-multipart==0.0.9

boto3>=1.34,<2.0
Pillow==10.4.0
numpy>=1.26,<3.0
opencv-python-headless>=4.10,<5.0
onnxruntime>=1.18,<2.0
